GUI/SeedFinder: Fix a crash on Windows when the recalculation ends
Again, another "why the heck Linux works?" type of crash.  Anyway, it was caused because Qt doesn't like to touch gui stuff outside of the gui thread so I changed it to an event. Also made it use a long for the value because I realized overflow is very possible.

+++ Source/GUI/SeedFinder/SeedFinderPassPage.cpp
@@ -58,7 +58,7 @@ void SeedFinderPassPage::setSeedFinderDone(const bool seedFinderDone)
   m_seedFinderDone = seedFinderDone;
 }
 
-void SeedFinderPassPage::setSeedFinderProgress(const int nbrSeedsSimulated)
+void SeedFinderPassPage::setSeedFinderProgress(const long nbrSeedsSimulated)
 {
   m_pbSeedFinder->setValue(nbrSeedsSimulated);
 }
